+++ src/app/modules/account/account.module.ts
@@ -19,11 +19,10 @@ import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { AccountRoutingModule } from "./account-routing.module";
 import { LoginComponent } from "./login/login.component";
-import { ClarityModule } from "@clr/angular";
 import { ReactiveFormsModule } from "@angular/forms";
 
 @NgModule({
-  imports: [CommonModule, AccountRoutingModule, ClarityModule, ReactiveFormsModule],
+  imports: [CommonModule, AccountRoutingModule, ReactiveFormsModule],
   declarations: [LoginComponent],
 })
 export class AccountModule {}