+++ src/app/modules/project/add/add.module.ts
@@ -18,13 +18,12 @@
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { ReactiveFormsModule } from "@angular/forms";
-import { ClarityModule } from "@clr/angular";
 import { AddRoutingModule } from "./add-routing.module";
 import { ManualComponent } from "./manual/manual.component";
 import { SourceComponent } from "./source/source.component";
 
 @NgModule({
   declarations: [SourceComponent, ManualComponent],
-  imports: [CommonModule, AddRoutingModule, ReactiveFormsModule, ClarityModule],
+  imports: [CommonModule, AddRoutingModule, ReactiveFormsModule],
 })
 export class AddModule {}