+++ src/app/modules/project/project.module.ts
@@ -17,13 +17,13 @@
 
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
-import { ClrAccordionModule } from "@clr/angular";
 import { DetailsComponent } from "./details/details.component";
 import { OverviewComponent } from "./overview/overview.component";
 import { ProjectRoutingModule } from "./project-routing.module";
+import { AccordionModule } from 'ngx-bootstrap/accordion';
 
 @NgModule({
   declarations: [OverviewComponent, DetailsComponent],
-  imports: [CommonModule, ProjectRoutingModule, ClrAccordionModule],
+    imports: [CommonModule, ProjectRoutingModule, AccordionModule.forRoot()],
 })
 export class ProjectModule {}