+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added this very changelog - [#69](https://github.com/DigitalExcellence/dex-backend/issues/71)
+- Added this very changelog - [#62](https://github.com/DigitalExcellence/dex-frontend/issues/62)
 - Added issue & pull request templates (bug & report) - [#11](https://github.com/DigitalExcellence/dex-backend/issues/11)
 
 ### Changed
+
+- Changed from Clarity framework to NGX Bootstrap - [#101](https://github.com/DigitalExcellence/dex-frontend/issues/101)
 
 ### Deprecated